Refactor start desktop layour
